feat: address and other pages config check

+++ database/migrations/2024_11_13_044123_add_column_to_products_table.php
@@ -11,18 +11,19 @@ return new class extends Migration
      */
     public function up(): void
     {
-        Schema::table('products', function (Blueprint $table) {
-            $table->string('expiry_date')->nullable();
-        });
+        if (Schema::hasTable('products') && !Schema::hasColumn('products', 'expiry_date')) {
+            Schema::table('products', function (Blueprint $table) {
+                $table->string('expiry_date')->nullable();
+            });
+        }
     }
 
-    /**
-     * Reverse the migrations.
-     */
     public function down(): void
     {
-        Schema::table('products', function (Blueprint $table) {
-            $table->dropColumn('expiry_date');
-        });
+        if (Schema::hasTable('products') && Schema::hasColumn('products', 'expiry_date')) {
+            Schema::table('products', function (Blueprint $table) {
+                $table->dropColumn('expiry_date');
+            });
+        }
     }
 };
